# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CERBERUS Commerce OS API
  version: "4.1-production-readiness"
  description: >
    Çekirdek route-handler sözleşmesi. Uçlar varsayılan olarak HttpOnly
    `cerberus_session` JWT çereziyle korunur. Açık uçlar `/api/auth/login`,
    `/api/health` ve `/api/health/ready` olarak security alanında ayrıca işaretlenir.
    Kimliksiz/geçersiz oturum 401, rol veya mağaza kapsamı ihlali 403, body
    doğrulama hatası 422, aşırı gövde 413 ve beklenmeyen sunucu hatası
    correlationId ile 500 döner. STORE_USER için mağaza kapsamı sunucuda zorlanır.
servers:
  - url: /
security:
  - cookieAuth: []
...
paths:
...
  /api/admin/users:
    get:
      tags: [admin]
      summary: Kullanıcı listesi (ADMIN)
      responses:
        "200": { description: "users (password_hash içermez)" }
        "401": { description: Oturum gerekli }
        "403": { description: Yetki yetersiz }
    post:
      tags: [admin]
      summary: Kullanıcı oluştur (ADMIN; parola bcrypt, min 12 kr.)
      responses:
        "200": { description: "message + user" }
        "422": { description: Doğrulama hatası }
    patch:
      tags: [admin]
      summary: Kullanıcı güncelle (ADMIN)
      responses:
        "200": { description: "message + user" }
        "422": { description: Doğrulama hatası }
  /api/admin/stores:
    get:
      tags: [admin]
      summary: Mağaza listesi + canlı istatistik
      description: STORE_USER yalnız kendi mağazasını; ADMIN/MANAGER tüm mağazaları görür. Hassas kart/e-posta alanları yalnız ADMIN'e açıktır.
      responses:
        "200": { description: Mağaza kapsamlı ve rol bazında maskelenmiş stores }
        "401": { description: Oturum gerekli }
    post:
      tags: [admin]
      summary: Mağaza oluştur (ADMIN/MANAGER)
      responses:
        "200": { description: "message + store" }
        "403": { description: Yetki yetersiz }
        "422": { description: Doğrulama hatası }
    patch:
      tags: [admin]
      summary: Mağaza güncelle (ADMIN/MANAGER)
      responses:
        "200": { description: "message + store" }
        "403": { description: Yetki yetersiz }
...
  /api/analytics/decision-support:
    get:
      tags: [analytics]
      summary: "Karar Destek Analitik — tek aggregate: KPI + trend + mağaza kıyas + PnL + fırsat + alarm"
      parameters:
        - { name: storeCode, in: query, schema: { type: string } }
        - { name: period, in: query, schema: { type: string, enum: [7d, 30d, 90d, all], default: 30d } }
      responses:
        "200": { description: "kpis, trend, storeComparison, opportunities, alerts, pnl" }
        "401": { description: Oturum gerekli }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: cerberus_session
      description: HttpOnly, imzalı ve süreli JWT oturum çerezi
...